ProductService 조회 기능 메서드 분리
