+++ app/src/main/java/com/codesoom/assignment/service/ProductService.java
@@ -6,10 +6,10 @@ import com.codesoom.assignment.repository.ProductJPARepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 import java.util.Optional;
 
 @Service
+@Transactional
 public class ProductService{
 
     private final ProductJPARepository repository;
@@ -19,36 +19,11 @@ public class ProductService{
     }
 
     /**
-     * 모든 엔티티를 반환합니다
-     *
-     * @return List<엔티티>
-     */
-    @Transactional(readOnly = true)
-    public List<Product> findAll(){
-        return (List<Product>) repository.findAll();
-    }
-
-    /**
-     * 식별자와 같은 엔티티를 찾아 반환합니다
-     *
-     * @param   id
-     * @return  id에 해당하는 자원
-     * @throws  ResourceNotFoundException
-     *          id에 해당하는 자원이 존재하지 않을 때 던져진다
-     */
-    @Transactional
-    public Product findById(Long id){
-        return repository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("상품이 존재하지 않습니다. id : " + id));
-    }
-
-    /**
      * 주어진 엔티티를 저장합니다
      *
      * @param   product
      * @return  저장된 product
      */
-    @Transactional
     public Product save(Product product){
         return repository.save(product);
     }
@@ -60,7 +35,6 @@ public class ProductService{
      * @param product 대체할 엔티티의 정보
      * @return 수정된 product
      */
-    @Transactional
     public Product update(Long id , Product product){
         Optional<Product> optionalProduct = repository.findById(id);
         optionalProduct
@@ -77,7 +51,6 @@ public class ProductService{
      * id에 해당하는 엔티티를 삭제한다
      * @param id 엔티티의 식별자
      */
-    @Transactional
     public void deleteById(Long id){
         repository.deleteById(id);
     }
